requirements.txt
# requirements.txt
flask
pandas
numpy
Pillow
psycopg2-binary
sqlalchemy
python-dotenv
matplotlib